[REFACTOR] make octa.hpp cleanly include-able
