+++ inexor/engine/octa.hpp
@@ -1,5 +1,19 @@
 // 6-directional octree heightfield map format
-//NO INCLUDE GUARD
+
+#pragma once
+
+#include "inexor/engine/world.hpp"
+#include "inexor/shared/geom.hpp"
+#include "inexor/shared/cube_vector.hpp"
+#include "inexor/shared/cube_types.hpp"
+#include "inexor/shared/ents.hpp"
+#include "inexor/network/SharedTree.hpp"
+
+#include <SDL.h>
+#include <SDL_opengl.h>
+
+#include <boost/algorithm/clamp.hpp> // TODO replace with std::clamp as soon as C++17 is our target.
+
 struct elementset
 {
     ushort texture, lmid, envmap;
@@ -210,6 +224,34 @@ struct cube
         uchar escaped;       // mask of which children have escaped merges
         uchar visible;       // visibility info for faces
     };
+};
+
+extern SharedVar<int> worldsize;
+
+/// Selection info: the marked area when editing.
+struct selinfo
+{
+    int corner;
+    int cx, cxs, cy, cys;
+    ivec o, s;
+    int grid, orient;
+    selinfo() : corner(0), cx(0), cxs(0), cy(0), cys(0), o(0, 0, 0), s(0, 0, 0), grid(8), orient(0) {}
+    int size() const { return s.x*s.y*s.z; }
+    int us(int d) const { return s[d]*grid; }
+    bool operator==(const selinfo &sel) const { return o==sel.o && s==sel.s && grid==sel.grid && orient==sel.orient; }
+    bool validate()
+    {
+        using boost::algorithm::clamp;
+        if(grid <= 0 || grid >= worldsize) return false;
+        if(o.x >= worldsize || o.y >= worldsize || o.z >= worldsize) return false;
+        if(o.x < 0) { s.x -= (grid - 1 - o.x)/grid; o.x = 0; }
+        if(o.y < 0) { s.y -= (grid - 1 - o.y)/grid; o.y = 0; }
+        if(o.z < 0) { s.z -= (grid - 1 - o.z)/grid; o.z = 0; }
+        s.x = clamp(s.x, 0, (worldsize - o.x)/grid);
+        s.y = clamp(s.y, 0, (worldsize - o.y)/grid);
+        s.z = clamp(s.z, 0, (worldsize - o.z)/grid);
+        return s.x > 0 && s.y > 0 && s.z > 0;
+    }
 };
 
 struct block3
