+++ inexor/shared/iengine.hpp
@@ -4,6 +4,8 @@
 #include "inexor/network/legacy/game_types.hpp"
 #include "inexor/util/legacy_time.hpp"
 
+#include "inexor/engine/octa.hpp"
+
 enum
 {
     MATF_INDEX_SHIFT  = 0,
@@ -52,35 +54,11 @@ extern bool isthirdperson();
 
 extern bool settexture(const char *name, int clamp = 0);
 
-extern SharedVar<int> worldsize;
 
 // octaedit
 
 enum { EDIT_FACE = 0, EDIT_TEX, EDIT_MAT, EDIT_FLIP, EDIT_COPY, EDIT_PASTE, EDIT_ROTATE, EDIT_REPLACE, EDIT_DELCUBE, EDIT_REMIP, EDIT_VSLOT, EDIT_UNDO, EDIT_REDO };
 
-struct selinfo
-{
-    int corner;
-    int cx, cxs, cy, cys;
-    ivec o, s;
-    int grid, orient;
-    selinfo() : corner(0), cx(0), cxs(0), cy(0), cys(0), o(0, 0, 0), s(0, 0, 0), grid(8), orient(0) {}
-    int size() const    { return s.x*s.y*s.z; }
-    int us(int d) const { return s[d]*grid; }
-    bool operator==(const selinfo &sel) const { return o==sel.o && s==sel.s && grid==sel.grid && orient==sel.orient; }
-    bool validate()
-    {
-        if(grid <= 0 || grid >= worldsize) return false;
-        if(o.x >= worldsize || o.y >= worldsize || o.z >= worldsize) return false;
-        if(o.x < 0) { s.x -= (grid - 1 - o.x)/grid; o.x = 0; } 
-        if(o.y < 0) { s.y -= (grid - 1 - o.y)/grid; o.y = 0; } 
-        if(o.z < 0) { s.z -= (grid - 1 - o.z)/grid; o.z = 0; } 
-        s.x = clamp(s.x, 0, (worldsize - o.x)/grid);
-        s.y = clamp(s.y, 0, (worldsize - o.y)/grid);
-        s.z = clamp(s.z, 0, (worldsize - o.z)/grid);
-        return s.x > 0 && s.y > 0 && s.z > 0;
-    }
-};
 
 struct editinfo;
 extern editinfo *localedit;
